moved CoordinatesVcf factory and mapper to core module
