+++ eva-accession-core/src/main/java/uk/ac/ebi/eva/accession/core/batch/io/CoordinatesVcfFactory.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package uk.ac.ebi.eva.accession.pipeline.batch.tasklets.reportCheck;
+package uk.ac.ebi.eva.accession.core.batch.io;
 
 import uk.ac.ebi.eva.commons.core.models.VariantCoreFields;
 import uk.ac.ebi.eva.commons.core.models.factories.VariantVcfFactory;
