+++ eva-accession-pipeline/src/test/java/uk/ac/ebi/eva/accession/pipeline/batch/tasklets/reportCheck/ReportCheckTaskletTest.java
@@ -27,6 +27,7 @@ import org.springframework.batch.item.NonTransientResourceException;
 import org.springframework.batch.item.ParseException;
 import org.springframework.batch.item.UnexpectedInputException;
 
+import uk.ac.ebi.eva.accession.core.batch.io.CoordinatesVcfLineMapper;
 import uk.ac.ebi.eva.accession.core.contig.ContigMapping;
 import uk.ac.ebi.eva.accession.core.contig.ContigSynonyms;
 import uk.ac.ebi.eva.commons.batch.io.AggregatedVcfReader;
